@@ -1,5 +1,10 @@
-module bystones.de/pwtrainer
+module github.com/bystones/pwtrainer
 
-go 1.12
+go 1.24.1
 
 require golang.org/x/crypto v0.19.0
+
+require (
+	golang.org/x/sys v0.17.0 // indirect
+	golang.org/x/term v0.17.0 // indirect
+)
